Fix Elixir `Blinky` to work on stm32
This should make elixir `Blinky` functional on all supported devices.
It will also work with any Pico device using an external LED on pin 2.
Poco-W uses can use the commented out PIN definition rather than the default to
use the onboard LED instead of connecting an external one, until we have a
programatic way to determine if the device is a regualr `pico` or a `picow`.

Signed-off-by: Winford <winford@object.stream>